Select driver based on Bargefile.
