+++ common/cmd.go
@@ -1,0 +1,15 @@
+package common
+
+import "io"
+
+// CmdInterface an interface wrapping the os/exec.Cmd struct. Makes command execution unit testable.
+type CmdInterface interface {
+	CombinedOutput() ([]byte, error)
+	Output() ([]byte, error)
+	Run() error
+	Start() error
+	StderrPipe() (io.ReadCloser, error)
+	StdinPipe() (io.WriteCloser, error)
+	StdoutPipe() (io.ReadCloser, error)
+	Wait() error
+}
